pyproject.toml
[project]
name = "moore-dataset-gen"
version = "0.1.0"
description = "YouTube → audio dataset pipeline for Mooré TTS (sidbi-ziri)"
requires-python = ">=3.10"

dependencies = [
    "yt-dlp",
    "boto3",
    "python-dotenv",
    "loguru",
    "tqdm",
    "pyyaml",
    "soundfile",
    "torch",
    "torchaudio",
    "transformers",
    "datasets",
    "huggingface_hub",
    "accelerate",
]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchaudio = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
